Better genome duplication section
I don't really know how to do with the reference that is currently
missing in this section. It is part of a wiki page that seems very
legit, and quite up to date. I don't want to refer to only that
though. There may be a review that I can use for that instead.